B Vector generator and legendre poly
Added in a function that takes in a value n and returns the vector of values of the integral from -1 to 1 of x^n. This will be used in the system of equations and the vanderman matrix that still needs to be constructed. I also added (and am still working on) a function that returns the roots of the Legendre polynomial as this will be needed in order to calculate the weights of the functions