Updated with Vandermonde matrix
Now we can ask for a Vandermonde matrix evaluated at the roots of the Legendre polynomial. The next step is to create the function that calculates the weights. Then we can form the summations equation to approximate the integral of a function. Looking forward to trying the full program on cosine